requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.txt
#
boto3==1.23.4
    # via -r requirements.txt
boto3-stubs[organizations]==1.23.0
    # via -r requirements.txt
botocore==1.26.10
    # via
    #   -r requirements.txt
    #   boto3
    #   s3transfer
botocore-stubs==1.27.4
    # via
    #   -r requirements.txt
    #   boto3-stubs
jmespath==1.0.0
    # via
    #   -r requirements.txt
    #   boto3
    #   botocore
mypy-boto3-organizations==1.23.0.post1
    # via
    #   -r requirements.txt
    #   boto3-stubs
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   botocore
pyyaml==6.0
    # via -r requirements.txt
s3transfer==0.5.2
    # via
    #   -r requirements.txt
    #   boto3
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
types-pyyaml==6.0.7
    # via -r requirements.txt
types-requests==2.27.30
    # via -r requirements.txt
types-setuptools==57.4.17
    # via -r requirements.txt
types-toml==0.10.7
    # via -r requirements.txt
types-urllib3==1.26.15
    # via
    #   -r requirements.txt
    #   types-requests
typing-extensions==4.2.0
    # via
    #   -r requirements.txt
    #   boto3-stubs
    #   botocore-stubs
    #   mypy-boto3-organizations
urllib3==1.26.9
    # via
    #   -r requirements.txt
    #   botocore
